VBoxVideo: add some explicit typecasts
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@37378 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/VMMDev/VMMDev.cpp
@@ -1486,7 +1486,7 @@ static DECLCALLBACK(int) vmmdevRequestHandler(PPDMDEVINS pDevIns, void *pvUser, 
             {
                 VMMDevVideoSetVisibleRegion *ptr = (VMMDevVideoSetVisibleRegion *)pRequestHeader;
 
-                if (pRequestHeader->size != sizeof(VMMDevVideoSetVisibleRegion) + ptr->cRect * sizeof(RTRECT) - sizeof(RTRECT))
+                if (pRequestHeader->size != ((uint64_t)sizeof(VMMDevVideoSetVisibleRegion)) + ptr->cRect * sizeof(RTRECT) - sizeof(RTRECT))
                 {
                     Log(("VMMDevReq_VideoSetVisibleRegion request size too small!!!\n"));
                     pRequestHeader->rc = VERR_INVALID_PARAMETER;
